requirements.txt
cassandra-driver
Flask
flask_restplus
flask_cassandra
pytest
tavern
